# docs/api/openapi-n1.json
{
  "components": {
    "schemas": {
...
      "PagamentoResponse": {
        "properties": {
          "expiraEm": {
            "format": "date-time",
            "type": "string"
          },
          "pagoEm": {
            "format": "date-time",
            "type": "string"
          },
          "pixCopiaECola": {
            "type": "string"
          },
          "provedor": {
            "enum": [
              "INTER",
              "SIMULADO"
            ],
            "type": "string"
          },
          "status": {
            "enum": [
              "PENDENTE",
              "PAGO",
              "EXPIRADO",
              "CANCELADO"
            ],
            "type": "string"
          },
          "txid": {
            "type": "string"
          },
          "valor": {
            "example": "80.00",
            "pattern": "^\\d+\\.\\d{2}$",
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Reserva de quadras esportivas com pagamento Pix.",
    "title": "SÃ³ mais uma â€” API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/dev/pagamentos/{txid}/confirmar": {
      "post": {
        "operationId": "confirmar",
        "parameters": [
          {
            "in": "path",
            "name": "txid",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Dev-Key",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PagamentoResponse"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {}
              }
            },
            "description": "ProblemDetail com codigo: TOKEN_INVALIDO"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {}
              }
            },
            "description": "ProblemDetail com codigo: ACESSO_NEGADO"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {}
              }
            },
            "description": "ProblemDetail com codigo: NAO_ENCONTRADO"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {}
              }
            },
            "description": "ProblemDetail com codigo: REGRA_NEGOCIO"
          },
          "502": {
            "content": {
              "application/problem+json": {
                "schema": {}
              }
            },
            "description": "ProblemDetail com codigo: PAGAMENTO_INDISPONIVEL"
          }
        },
        "summary": "Simula o pagamento da cobranÃ§a (exige X-Dev-Key e ser o cliente da reserva)",
        "tags": [
          "Dev"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "description": "desenvolvimento",
      "url": "http://localhost:8080"
    }
  ]
}